requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-cors-headers==4.4.0
django-jalali==7.1.0
django-nested-inline==0.4.6
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.4.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
